pyproject.toml
[tool.poetry]
name = "shape-recognition"
version = "0.1.0"
description = "C++ Shape Recognition with PyTorch (LibTorch) - macOS"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "shape_recognition", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8.1"
torch = "^2.2.0"
torchvision = "^0.17.0"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
matplotlib = "^3.7.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
train-model = "shape_recognition.train_model:main"
test-setup = "shape_recognition.test_setup:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 